Adds loading of redux store with proper force-updating react components.
