+++ src/logger.middleware.js
@@ -1,0 +1,9 @@
+
+export const logger = (store) => (next) => (action) => {
+  console.group(action.type)
+    console.log('The action: ', action)
+    const result = next(action)
+    console.log('The new state: ', store.getState())
+  console.groupEnd()
+  return result
+}
