+++ src/store.js
@@ -1,18 +1,10 @@
 import {applyMiddleware, combineReducers, createStore} from 'redux';
 import {users} from "./users.reducer";
+import {loading} from "./loading.reducer";
+import {logger} from './logger.middleware';
 
-
-
-const logger = (store) => (next) => (action) => {
-  console.group(action.type)
-    console.log('The action: ', action)
-    const result = next(action)
-    console.log('The new state: ', store.getState())
-  console.groupEnd()
-  return result
-}
 const store = createStore(combineReducers({
-    users,
+    users, loading,
 }), applyMiddleware(logger))
 
 export default store;
